cambio de tipo de dato en precio unitario

+++ Domain/Entities/Producto.cs
@@ -5,7 +5,7 @@ namespace Domain.Entities
     {
         public int Id { get; set; }
         required public string Nombre { get; set; }
-        required public double PrecioUnitario { get; set; }
+        required public decimal PrecioUnitario { get; set; }
 
         required public string Descripcion { get; set; }
